Rewrite ChangeCommand.
New version of ChangeCommand.do_execute that does not use the
'manager' or the 'target_builder' - i.e., these lines were removed:
    manager.target_builder.spec_collector = options
    manager.target_builder.set_edit_mode
    manager.update_targets(options)
and replaced with code that, based on the arguments in
'request.arguments', retrieves the specified target (via handle)
and makes the specified changes. It uses the 'new_spec' method that
was in the AddCommand code. Thus 'new_spec' was moved into the
parent class, WorkCommand.