remove_descendant DB bug fix; etc.
ChangeHandleCommand: Gemini-cli's code in do_execute was not quite
right. Fixed/improved it.

RedisBasedDataManager: Fixed a problem with 'store_target'.

STodoTarget.remove_descendant: Moved:
    self.last_removed_descendant = child
to after the 'update'. It was causing a "nested too deep" error
when the STodoTarget was saved in the database.

Took care of some code I missed when getting rid of use of the
legacy 'manager'.