Apparent DB lost-data problem fixed
Fixed the problem with the report_command_e2e test failing due
to apparently lost objects in redis - i.e., objects that the
server said were inserted (with a success response) but were
not showing up in report queries. The problem was due to
a "left-over" user and app_name from previous executions. The
user and app_name were not being set at the point they should
have been, which is just before any database insertions or
queries. The fix, in WorkCommand.execute:
+    database.set_appname_and_user(request.app_name, request.user_id)

Removed the 'sleep's in report_command_e2e.sh, since they appear
to actually not be needed.