Test fixes; changes: NullSTodoGit, database, notification
NotificationService: A few improvements to 'process' to get it
to work correctly. Also, made 'dirty' and 'dirty=' public, since
STodoTarget calls 'dirty='.

RedisBasedDataManager: Changed 'def initialize' to not require
a non-nil user and appname - set_appname_and_user can be used
to set these values.

Changes to configuration.rb and configtools.rb to "plug in" the
new NullSTodoGit, which provides an option to replace STodoGit to
provide a "dummy" git that pretends to provide the STodoGit
operations, but does nothing (the purpose being to bypass the
use of git, which causes lock, etc. problems with the new
architecture). And changed STodoGit to get rid of the pre/post-
condition checks, which were causing problems when STodoGit was
used.

Test fixes to work with new changes in the new architecture -
commands, etc: add-change-delete-etc-101, clone-101, clone_command_e2e,
tools (test utilities).